% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_px.R
\name{ggsave_px}
\alias{ggsave_px}
\alias{ggsave_1080p}
\alias{ggsave_1440p}
\alias{ggsave_2160p}
\title{Save ggplot in px}
\usage{
ggsave_px(filename, plot, height, width, aspect, dpi, type)

ggsave_1080p(filename, plot, aspect = "wide", dpi = 300, type = "cairo")

ggsave_1440p(filename, plot, aspect = "wide", dpi = 300, type = "cairo")

ggsave_2160p(filename, plot, aspect = "wide", dpi = 300, type = "cairo")
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{height}{Plot height in pixels}

\item{width}{Plot width in pixels}

\item{aspect}{wide (16:9), tall (9:16), square (1:1)}

\item{dpi}{Plot resolution.}

\item{type}{type}
}
\description{
Save ggplot in px
}
\section{Functions}{
\itemize{
\item \code{ggsave_1080p()}: Save in FHD (1080p)

\item \code{ggsave_1440p()}: Save in QHD (1440p)

\item \code{ggsave_2160p()}: Save in 4k (2160p)

}}
