% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reto.R
\name{theme_reto}
\alias{theme_reto}
\title{The reto theme}
\usage{
theme_reto(variant = "light", base_size = 18, font = "CMU Sans Serif")
}
\arguments{
\item{variant}{the theme variant (light, dark)}

\item{base_size}{the font size}

\item{font}{the font}
}
\value{
a ggplot theme
}
\description{
The reto theme
}
