% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reto.R
\name{theme_reto_}
\alias{theme_reto_}
\title{reto theme base function}
\usage{
theme_reto_(
  base_size,
  font,
  background_color,
  canvas_primary_color,
  canvas_secondary_color,
  font_primary_color,
  font_secondary_color
)
}
\arguments{
\item{base_size}{font size}

\item{font}{font}

\item{background_color}{the background color of the plot}

\item{canvas_primary_color}{applies to legend and plot area}

\item{canvas_secondary_color}{applies to strip (facet-title)}

\item{font_primary_color}{applies to most text}

\item{font_secondary_color}{applies to subtitle etc.}
}
\value{
a ggplot theme
}
\description{
reto theme base function
}
