% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_reto.R
\name{theme_reto}
\alias{theme_reto}
\title{Custom Theme Reto}
\usage{
theme_reto(base_size = 18, font = "CMU Sans Serif")
}
\arguments{
\item{base_size}{font size}

\item{font}{font}
}
\value{
plot obj
}
\description{
Custom Theme Reto
}
